requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
stripe==6.7.0
python-multipart==0.0.6
jinja2==3.1.2
